requirements.txt
openai>=1.0.0
tqdm>=4.60.0
Pillow>=9.0.0
opencv-python>=4.5.0
